Update to GNOME 47 runtime
python3-build-dependencies.json was generated with:

    flatpak-pip-generator \
        --runtime org.gnome.Sdk//47 \
        -r build-requirements.txt \
        --cleanup all \
        --output python3-build-dependencies

python3-requirements.json was generated with:

    flatpak-pip-generator \
        --runtime org.gnome.Sdk//47 \
        -r requirements.txt

The generated manifest was then manually adjusted to fix a build issue with
matplotlib. By default, matplotlib tries to download and build freetype and
qhull during its build process. This fails (by hanging indefinitely!) in
the Flatpak build sandbox, which does not have network access. To solve this,
build qhull as part of the main manifest, then adjust the `pip3 install`
incantation in the generated manifest to pass the necessary options to
matplotlib's `meson setup` to use the "system" versions of this library.